requirements.txt
kuksa-client
requests
eclipse-zenoh
flask
